Add support for Timeout, Verbose  and more commands
When the provided timeout is not 0 the commands will return None if the timeout occurs, instead of just waiting forever. The default remains at wait forever for compatiblity

When Verbose is set to True, more information will be displayed, including the command that is sent.

Added multiline support for the memory and program commands.

Partial support for the screen command, just reading important status from it

Added